// reiDoCangaco/src/app/models/fornecedorPf.model.ts
export interface FornecedorPf {
    idFornecedorPF?: string;
    nome: string;
    nickName: string;
    sexo?: string;
    cpf?: string;
    dataNascimento: string;
    telefone: string;
    celular: string;
    email: string;
    endereco: string;
    complemento?: string;
    bairro: string;
    cep: string;
    cidade: string;
    estado: string;
    pais: string;
    status?: string;
    flagAtivo?: boolean;
}