// reiDoCangaco/src/app/models/fornecedorPj.model.ts
export interface FornecedorPj {

    idFornecedorPJ?: string;
    razaoSocial: string;
    nomeFantasia: string;
    nickName?: string;
    cnpj: string;
    telefone: string;
    celular: string;
    email: string;
    endereco: string;
    complemento: string;
    bairro: string;
    cep: string;
    cidade: string;
    estado: string;
    pais: string;
    status?: string;
    flagAtivo?: boolean;
}